fix(www): revalidate plant sorts for seed changes
